added regex validation on amount, some methods to mixins to cleanup code
